requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
passlib
bcrypt
python-jose[cryptography]
python-dotenv
black
python-multipart
boto3
uuid
pydantic_computed
